cut_tiles.py added
run as a standalone program with command line parameters, it can slice one image into tiles with potentially 7 zoom levels (0 to 6).

run without command line parameters, it assumes that the full maps are png files in the full_maps directory, and it proceeds to slice them into tiles and to organize them as subdirectories of dist (with a structure that is easily managed by the main il2missionplanner.link application).